Move keyword validation to common file
